requirements.txt
botorch==0.10.0
flatten_dict==0.4.2
gpytorch==1.11
matplotlib==3.4.3
numpy==1.22.4
omegaconf==2.3.0
pandas==2.2.2
python-dotenv==1.1.0
scikit_learn==1.5.0
scipy==1.13.0
sobol_seq==0.2.0
torch==2.3.0+cu118
torchrl==0.4.0
tqdm==4.62.3
